Crate pallet_block_producer_fees

Pallet to store Block Producer Fees settings that are relevant to rewards payments.

Margin fee is percent of block rewards that will be paid to the block producer before distributing the rest of rewards to his stakers. Precision of the margin fee setting is bounded to 1/100 of a percent.

The margin fees are stored together with the slot at which change occurred, so this data can be exposed to rewards calculation.

Log of changes per account is bounded. The oldest entries are dropped when new ones are added. Intention is to discourage users from too frequent changes and there is an assumption that rewards calculation algorithm will account for it.

Re-exports

pub use weights::WeightInfo;

pub use pallet::*;

Modules

pallet

The pallet module in each FRAME pallet hosts the most important items needed to construct this pallet.

weights

Autogenerated weights for pallet_block_producer_fees